+++ src/tasks/sourceserifpro.rs
@@ -1,0 +1,26 @@
+use crate::lib::{
+    ghrafont::GhraFont,
+    task::{self, Task},
+};
+
+pub fn task() -> Task {
+    Task {
+        name: String::from("sourceserifpro"),
+        sync,
+        update,
+    }
+}
+
+const GHRA_FONT: GhraFont = GhraFont {
+    asset_re: r"^source-serif-pro-.*\.zip$",
+    font_suffix: ".otf",
+    repo: ("adobe-fonts", "source-serif-pro"),
+};
+
+fn sync() -> task::Result {
+    GHRA_FONT.sync()
+}
+
+fn update() -> task::Result {
+    GHRA_FONT.update()
+}