+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- plex: keep [Plex (font)](https://github.com/IBM/plex) `sync`ed and `update`d
+
+- sourcesanspro: keep [Source Sans Pro (font)](https://github.com/adobe-fonts/source-sans-pro) `sync`ed and `update`d
+
+- sourceserifpro: keep [Source Serif Pro (font)](https://github.com/adobe-fonts/source-serif-pro) `sync`ed and `update`d
+
 ## [0.31.0] - 2019-04-19
 
 ### Added